add sam

+++ model/sam/mask_decoder.h
@@ -6,7 +6,7 @@ class MaskDecoder : public CvModel {
     MaskDecoder(const std::string &model_path, const std::string framework_type);
     MaskDecoder(const std::string &yaml_file);
     virtual ~MaskDecoder();
-    void forward(const IOTensor &features, const std::vector<cv::Point2i> &image_point_coords,
+    void forward(const IOTensor &features, const std::vector<cv::Point2f> &image_point_coords,
                  const std::vector<float> &image_point_labels, cv::Mat &low_res_mask);
 
    private:
